requirements.txt
langchain>=1.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
langfuse>=2.50.0
openai>=1.54.0
chromadb>=0.5.0
tiktoken>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pytest>=7.4.0
pytest-mock>=3.12.0
